Enum relay_sampling::condition::RuleCondition

pub enum RuleCondition {
    Eq(EqCondition),
    Gte(GteCondition),
    Lte(LteCondition),
    Gt(GtCondition),
    Lt(LtCondition),
    Glob(GlobCondition),
    Or(OrCondition),
    And(AndCondition),
    Not(NotCondition),
    Unsupported,
}

A condition from a sampling rule.

Variants

Eq(EqCondition)

A condition that compares values for equality.

Gte(GteCondition)

A condition that applies >=.

Lte(LteCondition)

A condition that applies <=.

Gt(GtCondition)

A condition that applies >.

Lt(LtCondition)

A condition that applies <.

Glob(GlobCondition)

A condition that uses glob matching.

Or(OrCondition)

Combines multiple conditions using logical OR.

And(AndCondition)

Combines multiple conditions using logical AND.

Not(NotCondition)

Applies logical NOT to a condition.

Unsupported

An unsupported condition for future compatibility.

Implementations

impl RuleCondition

pub fn all() -> Self

Returns a condition that matches everything.

pub fn supported(&self) -> bool

Checks if Relay supports this condition (in other words if the condition had any unknown configuration which was serialized as “Unsupported” (because the configuration is either faulty or was created for a newer relay that supports some other condition types)

pub fn matches<T>(&self, value: &T) -> boolwhere T: Getter,

Returns true if the rule matches the given value instance.