Struct relay_sampling::config::SamplingConfig

pub struct SamplingConfig {
    pub rules: Vec<SamplingRule>,
    pub rules_v2: Vec<SamplingRule>,
    pub mode: SamplingMode,
}

Represents the dynamic sampling configuration available to a project.

Note: This comes from the organization data

Fields
rules: Vec<SamplingRule>

The ordered sampling rules for the project.

This field will remain here to serve only for old customer Relays to which we will forward the sampling config. The idea is that those Relays will get the old rules as empty array, which will result in them not sampling and forwarding sampling decisions to upstream Relays.

rules_v2: Vec<SamplingRule>

The ordered sampling rules v2 for the project.

mode: SamplingMode

Defines which population of items a dynamic sample rate applies to.

Implementations

impl SamplingConfig

pub fn unsupported(&self) -> bool

Returns true if any of the rules in this configuration is unsupported.

pub fn filter_rules( &self, rule_type: RuleType ) -> impl Iterator<Item = &SamplingRule>

Filters the sampling rules by the given RuleType.