relay_sampling::dsc

Struct DynamicSamplingContext

pub struct DynamicSamplingContext {
    pub trace_id: TraceId,
    pub public_key: ProjectKey,
    pub release: Option<String>,
    pub environment: Option<String>,
    pub transaction: Option<String>,
    pub sample_rate: Option<f64>,
    pub user: TraceUserContext,
    pub replay_id: Option<Uuid>,
    pub sampled: Option<bool>,
    pub other: BTreeMap<String, Value>,
}

DynamicSamplingContext created by the first Sentry SDK in the call chain.

Because SDKs need to funnel this data through the baggage header, this needs to be representable as HashMap<String, String>, meaning no nested dictionaries/objects, arrays or other non-string values.

Fields
trace_id: TraceId

ID created by clients to represent the current call flow.

public_key: ProjectKey

The project key.

release: Option<String>

The release.

environment: Option<String>

The environment.

transaction: Option<String>

In the transaction-based model, this is the name of the transaction extracted from the transaction field in the starting transaction and set on transaction start, or via scope.transaction.

In the spans-only model, this is the segment name for the segment that started the trace.

sample_rate: Option<f64>

The sample rate with which this trace was sampled in the client. This is a float between 0.0 and 1.0.

user: TraceUserContext

The user specific identifier (e.g. a user segment, or similar created by the SDK from the user object).

replay_id: Option<Uuid>

If the event occurred during a session replay, the associated replay_id is added to the DSC.

sampled: Option<bool>

Set to true if the transaction starting the trace has been kept by client side sampling.

other: BTreeMap<String, Value>

Additional arbitrary fields for forwards compatibility.