Enum relay_sampling::evaluation::Evaluation

pub enum Evaluation {
    Continue(SamplingEvaluator),
    Matched(SamplingMatch),
}

State of the SamplingEvaluator’s rule matching.

Enforces a pattern to avoid matching on more rules if a match has already been found.

Variants

Continue(SamplingEvaluator)

Matching has not completed and more rules can be evaluated.

This can happen either if no active rules match the provided data, or if the matched rules are factors that require a final sampling value. In this case, the returned evaluator stores the state of the matched rules as well as the accumulated sampling factor and can be used to evaluate more rules.

If evaluation returns Continue and there are no more rules to match, this should be considered as “no match”.

Matched(SamplingMatch)

One or more rules have matched.

Implementations

impl Evaluation

pub fn is_match(&self) -> bool

Returns true if a rule has matched.

pub fn is_no_match(&self) -> bool

Returns true if no rule have matched.