Function relay_sampling::merge_rules_from_configs

pub fn merge_rules_from_configs<'a>(
    sampling_config: &'a SamplingConfig,
    root_sampling_config: Option<&'a SamplingConfig>
) -> impl Iterator<Item = &'a SamplingRule>

Returns an iterator of references that chains together and merges rules.

The chaining logic will take all the non-trace rules from the project and all the trace/unsupported rules from the root project and concatenate them.