Crate relay_sampling

Sampling logic for performing sampling decisions of incoming events.

In order to allow Sentry to offer performance at scale, Relay extracts key metrics from all transactions, but only forwards a random sample of raw transaction payloads to the upstream. What exact percentage is sampled is determined by dynamic sampling rules, and depends on the project, the environment, the transaction name, etc.

In order to determine the sample rate, Relay uses a SamplingConfig which contains a set of SamplingRules that are matched against the incoming [Event] or DynamicSamplingContext.

Trace and Transaction Sampling

Relay samples both transactions looking at [Event] and traces looking at DynamicSamplingContext:

Trace sampling: ensures that either all transactions of a trace are sampled or none.

Transaction sampling: does not guarantee complete traces and instead applies to individual transactions.

Components

The sampling system implemented in Relay is composed of the following components:

DynamicSamplingContext: a struct that contains the trace information.

FieldValueProvider: an abstraction implemented by [Event] and DynamicSamplingContext to expose fields that are read during matching.

SamplingRule: a rule that is matched against [Event] or DynamicSamplingContext that can contain a RuleCondition for expressing predicates on the incoming payload.

SamplingMatch: the result of the matching of one or more SamplingRule.

How It Works

The incoming [Event] and optionally DynamicSamplingContext are received by Relay.

Relay fetches the SamplingConfig of the project to which the [Event] belongs and (if exists) the SamplingConfig of the root project of the trace.

The SamplingConfigs are merged together and the matching algorithm in SamplingMatch::match_against_rules is executed.

The sampling algorithm will go over each SamplingRule and compute either a factor or sample rate based on the SamplingValue of the rule.

The SamplingMatch is finally returned containing the final sample_rate and some additional data that will be used in relay_server to perform the sampling decision.

Sampling Determinism

The concept of determinism is extremely important for sampling. We want to be able to make the a deterministic sampling decision for a wide variety of reasons, including:

Across a chain of Relays (e.g., we don’t want to drop an event that was retained by a previous Relay and vice-versa).

Across transactions of the same trace (e.g., we want to be able to sample all the transactions of the same trace, even though some exceptions apply).

In order to perform deterministic sampling, we use the id of the event or trace as the seed for the random number generator (e.g., all transactions with the same trace id will have the same random number being generated). Since we allow the matching of both transaction and trace rules, we might end up in cases in which we perform inconsistent trace sampling but this is something we decided to live with as long as there are no big implications on the product.

Examples

SamplingConfig

{
  "rules": [],
  "rulesV2": [
    {
      "condition": {
        "op": "and",
        "inner": [
          {
            "op": "glob",
            "name": "releases",
            "value": [
              "1.1.1",
              "1.1.2"
            ]
          }
        ]
      },
      "samplingValue": {
        "type": "sampleRate",
        "value": 0.7
      },
      "type": "transaction",
      "id": 2
    },
    {
      "condition": {
        "op": "and",
        "inner": [
          {
            "op": "eq",
            "name": "transaction",
            "value": "/foo"
          }
        ]
      },
      "samplingValue": {
        "type": "factor",
        "value": 5.0
      },
      "type": "trace",
      "id": 1,
      "timeRange": {
        "start": "2022-10-10T00:00:00.000000Z",
        "end": "2022-10-20T00:00:00.000000Z"
      }
    }
  ],
  "mode": "total"
}

DynamicSamplingContext

{
  "trace_id": "00000000-0000-0000-0000-000000000000",
  "public_key": "abd0f232775f45feab79864e580d160b",
  "release": "1.0",
  "environment": "dev",
  "transaction": "/foo",
  "sample_rate": "0.5",
  "user_id": "some-id",
  "user_segment": "all",
  "user": null
}

Structs

ActiveRule

A sampling rule that has been successfully matched and that contains all the required data to return the sample rate.

AndCondition

And condition combinator.

DynamicSamplingContext

DynamicSamplingContext created by the first Sentry SDK in the call chain.

EqCondOptions

A condition that checks the values using the equality operator.

EqCondition

A condition that checks for equality

GlobCondition

A condition that uses glob matching.

GtCondition

GteCondition

LtCondition

LteCondition

MatchedRuleIds

Represents a list of rule ids which is used for outcomes.

NotCondition

Not condition combinator.

OrCondition

Or condition combinator.

RuleId

The id of the SamplingRule.

SamplingConfig

Represents the dynamic sampling configuration available to a project.

SamplingMatch

Represents the specification for sampling an incoming event.

SamplingRule

A sampling rule as it is deserialized from the project configuration.

TimeRange

A range of time.

TraceUserContext

The User related information in the trace context

Enums

DecayingFunction

A decaying function definition.

RuleCondition

A condition from a sampling rule.

RuleType

Defines the type of dynamic rule, i.e. to which type of events it will be applied and how.

SamplingMode

Defines which population of items a dynamic sample rate applies to.

SamplingValue

A sampling strategy definition.

Traits

FieldValueProvider

Trait implemented by providers of fields (Events and Trace Contexts).

Functions

deserialize_bool_option

merge_configs_and_match

Gets the sampling match result by creating the merged configuration and matching it against the sampling configuration.

merge_rules_from_configs

Returns an iterator of references that chains together and merges rules.

pseudo_random_from_uuid

Generates a pseudo random number by seeding the generator with the given id.