relay_sampling

Module dsc

Contextual information stored on traces.

DynamicSamplingContext (DSC) contains properties associated to traces that are shared between individual data submissions to Sentry. These properties are supposed to not change during the lifetime of the trace.

The information in the DSC is used to compute a deterministic sampling decision without access to all individual data in the trace.

Structs

DynamicSamplingContext

DynamicSamplingContext created by the first Sentry SDK in the call chain.

TraceUserContext

User-related information in a DynamicSamplingContext.