Trait relay_sampling::FieldValueProvider

pub trait FieldValueProvider {
    // Required methods
    fn get_value(&self, path: &str) -> Value;
    fn get_custom_operator(
        name: &str
    ) -> fn(condition: &CustomCondition, slf: &Self, ip_addr: Option<IpAddr>) -> bool;
}

Trait implemented by providers of fields (Events and Trace Contexts).

The fields will be used by rules to check if they apply.

Required Methods

fn get_value(&self, path: &str) -> Value

gets the value of a field

fn get_custom_operator( name: &str ) -> fn(condition: &CustomCondition, slf: &Self, ip_addr: Option<IpAddr>) -> bool

returns a filtering function for custom operators. The function returned takes the provider and a condition definition and returns a match result